# executor, types: fix bug of insert into unsigned float/double