@coocood 2018-06-29
/run-common-test
/run-integration-common-test